refactor: :recycle: set deques from to
