+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/11 16:10:56 by youkim            #+#    #+#             */
-/*   Updated: 2021/12/12 15:38:21 by youkim           ###   ########.fr       */
+/*   Updated: 2021/12/12 20:00:26 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,13 @@
 # define PUSH_SWAP_H
 
 # include "libft.h"
+# include "structure.h"
 # include "engine.h"
 # include "opers.h"
 
+# define operl(engine, op) 				\
+		do { oper(engine, op);			\
+			system("clear");			\
+			engine_visualize(engine);	\
+			usleep(400000); } while (0);
 #endif
